feat!: Assume authors with same email are the same person
Previously authors with the same email, but a different spelling of the
name were considered different persons. We now use the email to identify
an author.
`gitSuggestReviewer` now returns only one entry per email. The first
name that was processed is used in `ReviewerStats`.

This would lead to one single entry for 'Peter Pan' now, not two:

```
Author: Peter Pan <peter@pan.de>
Author: Peterchen Pan <peter@pan.de>
```